MOVEONLY: Move to consensus/blockverify.cpp:
(declarations to consensus/consensus.h)

from main.cpp:
-CheckBlockHeader
-ContextualCheckBlockHeader
-IsSuperMajority

from pow.cpp:
-CalculateNextWorkRequired
-CheckProofOfWork
-GetNextWorkRequired
